overlap change type to int

+++ darbukaToneIdentification/views.py
@@ -209,7 +209,7 @@ def userIdentification(request):
 
   mfcc_parameter = mfcc_parameters.objects.all()[0]
   frameLength = float(mfcc_parameter.frame_length)
-  overlap = float(mfcc_parameter.overlap)
+  overlap = int(mfcc_parameter.overlap)
   mfccCoefficients = int(mfcc_parameter.mfcc_coefficient)
 
   context = {}
